requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
google-auth>=2.17.0
google-api-python-client>=2.100.0
gspread>=5.10.0
python-dotenv>=1.0.0